コマンドをハンダラとしてわけつつ上手く動作しない問題を修正
